withrestc1/requirements.txt
argon2-cffi==18.3.0
attrs==18.1.0
bcrypt==3.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
cx-Oracle==6.4.1
Django==1.11.16
django-filter==2.0.0
django-taggit==0.23.0
djangorestframework==3.9.0
Faker==0.9.0
Flask==1.0.2
httpie==1.0.2
idna==2.8
Interface==2.11.1
itsdangerous==0.24
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
more-itertools==4.1.0
mysqlclient==1.3.12
pluggy==0.6.0
py==1.5.3
pycparser==2.18
pygame==1.9.3
Pygments==2.3.0
PyMySQL==0.8.1
pytest==3.5.1
pytest-html==1.18.0
pytest-metadata==1.7.0
pytest-ordering==0.5
python-dateutil==2.7.3
python-interface==1.4.0
pytz==2018.7
requests==2.21.0
selenium==3.11.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
Werkzeug==0.14.1
zope.event==4.3.0
zope.interface==4.5.0
zope.schema==4.5.0
